feat: refactor contests leaderboards
